app/server/requirements.txt
pandas
fastapi
python-dotenv
requests
uvicorn